requirements.txt
absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
category-encoders==2.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
Django==4.0.4
docopt==0.6.2
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-Table==0.5.0
Flask-WTF==1.0.1
flatbuffers==1.12
fonttools==4.37.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.0
nltk==3.7
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
patsy==0.5.2
pdfkit==1.0.0
Pillow==9.2.0
pipreqs==0.4.11
preprocessor==1.1.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
statsmodels==0.13.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yarg==0.1.9
